requirements.txt
streamlit
sentence-transformers
transformers
faiss-cpu>=1.6.5
torch
pypdf
pandas